expenses-manager-api/requirements.txt
flask==3.0.2
psycopg2=2.9.9
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1

